refactoring : payment service 나눔 - 두 개의 도메인에 대해서 service를 나누었습니다 두 service는 나중에 인터페이스를 상속받아서 구현하겠습니다
